Initialize MikroOrm with a PostgreSQL database
This commit covers the initialization of a MicroOrm instance and
connecting to a PostgreSQL database, creating a Post entity, setting
up the MicroOrm cli, creating and executing migrations to create a
Post table using the cli.

Added heavy comments for narration, and easy navigation.

Initialize MikroOrm with a PostgreSQL database

This commit covers the initialization of a MicroOrm instance and
connecting to a PostgreSQL database, creating a Post entity, setting
up the MicroOrm cli, creating and executing migrations to create a
Post table using the cli.

Added heavy comments for narration, and easy navigation.

forgot to commit these guys